# Backend/openapi-v1.yaml
openapi: 3.0.3
info:
  title: GahamBani API
  version: 1.0.0
  description: Documentation de l'API GahamBani. Le Domaine 1 couvre l'authentification
    et la gestion du compte utilisateur.
paths:
...
  /api/v1/notifications/{id}/read/:
    patch:
      operationId: api_v1_notifications_read_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this notification.
        required: true
      tags:
      - Notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
        '400':
          description: Filtre invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '403':
          description: Compte ou propriété insuffisante.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '404':
          description: Notification inexistante ou inaccessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '405':
          description: Méthode désactivée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/notifications/read-all/:
    patch:
      operationId: api_v1_notifications_read_all_partial_update
      tags:
      - Notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadAllResult'
          description: ''
        '400':
          description: Filtre invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '403':
          description: Compte ou propriété insuffisante.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '404':
          description: Notification inexistante ou inaccessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '405':
          description: Méthode désactivée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/notifications/unread-count/:
    get:
      operationId: api_v1_notifications_unread_count_retrieve
      tags:
      - Notifications
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCount'
          description: ''
        '400':
          description: Filtre invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '403':
          description: Compte ou propriété insuffisante.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '404':
          description: Notification inexistante ou inaccessible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '405':
          description: Méthode désactivée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/NotificationTypeEnum'
          readOnly: true
        priority:
          allOf:
          - $ref: '#/components/schemas/PriorityEnum'
          readOnly: true
        title:
          type: string
          readOnly: true
        message:
          type: string
          readOnly: true
        is_read:
          type: boolean
          readOnly: true
        read_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        source_domain:
          type: string
          readOnly: true
        source_type:
          type: string
          readOnly: true
        source_id:
          type: string
          format: uuid
          readOnly: true
        event_code:
          type: string
          readOnly: true
        metadata:
          readOnly: true
      required:
      - created_at
      - event_code
      - id
      - is_read
      - message
      - metadata
      - priority
      - read_at
      - source_domain
      - source_id
      - source_type
      - title
      - type
    NotificationTypeEnum:
      enum:
      - MEDICAL_ALERT_CREATED
      - ALERT_ACKNOWLEDGED
      - ALERT_RESOLVED
      - ALERT_DISMISSED
      - SYSTEM
      type: string
      description: |-
        * `MEDICAL_ALERT_CREATED` - Medical alert created
        * `ALERT_ACKNOWLEDGED` - Alert acknowledged
        * `ALERT_RESOLVED` - Alert resolved
        * `ALERT_DISMISSED` - Alert dismissed
        * `SYSTEM` - System
...
    PriorityEnum:
      enum:
      - LOW
      - NORMAL
      - HIGH
      - CRITICAL
      type: string
      description: |-
        * `LOW` - Low
        * `NORMAL` - Normal
        * `HIGH` - High
        * `CRITICAL` - Critical
...
    ReadAllResult:
      type: object
      properties:
        updated_count:
          type: integer
          readOnly: true
      required:
      - updated_count
...
    UnreadCount:
      type: object
      properties:
        unread_count:
          type: integer
          readOnly: true
      required:
      - unread_count
...
    V1Error:
      type: object
      required:
      - code
      - detail
      properties:
        code:
          type: string
          example: validation_error
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        retry_after:
          type: integer
          minimum: 0
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT